upcoming : [DI-19807] - Modified the Polling Inteval filtering logic
