+++ packages/manager/src/features/CloudPulse/Alerts/AlertLanding/RecentActivity.tsx
@@ -3,8 +3,6 @@ import * as React from 'react';
 import { Paper } from 'src/components/Paper';
 import { Typography } from 'src/components/Typography';
 
-import CreateAlertDefinitionDrawer from '../CreateAlert';
-
 export const RecentActivity = React.memo(() => {
 
   return (
