--- packages/manager/src/features/CloudPulse/Alerts/CreateAlert/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CreateAlertDefinitionDrawer';
