+++ packages/manager/src/features/CloudPulse/Alerts/CreateAlert/shared/ResourceMultiSelect.tsx
@@ -1,13 +1,8 @@
-/* eslint-disable no-console */
 import { useFormikContext } from 'formik';
 import * as React from 'react';
 
 import { Autocomplete } from 'src/components/Autocomplete/Autocomplete';
-import {
-  useLinodeResourcesQuery,
-  useLoadBalancerResourcesQuery,
-} from 'src/queries/cloudpulse/resources';
-import { getAllDatabases, useDatabaseEnginesQuery } from 'src/queries/databases';
+import { useLinodeResourcesQuery } from 'src/queries/cloudpulse/resources';
 
 interface CloudViewResourceSelectProps {
   cluster: boolean;
@@ -24,9 +19,7 @@ export const CloudViewMultiResourceSelect = (
   const resourceOptions: any = {};
 
   const formik = useFormikContext();
-  const values = formik.getFieldProps(props.name).value;
   const [selectedResource, setResource] = React.useState<any>([]);
-  // const [resourceInputValue, setResourceInputValue] = React.useState<any>('');
   const filterResourcesByRegion = (resourcesList: any[]) => {
     return resourcesList?.filter((resource: any) => {
       if (props.region == undefined) {
@@ -43,7 +36,6 @@ export const CloudViewMultiResourceSelect = (
   };
 
   const getResourceList = () => {
-    // console.log(resourceOptions[props.resourceType!]);
     return props.resourceType && resourceOptions[props.resourceType]
       ? filterResourcesByRegion(resourceOptions[props.resourceType]?.data)
       : [];
@@ -52,13 +44,6 @@ export const CloudViewMultiResourceSelect = (
   ({ data: resourceOptions['linode'] } = useLinodeResourcesQuery(
     props.resourceType === 'linode'
   ));
-  // ({ data: resourceOptions['db'] } = useDatabaseEnginesQuery(
-  //   props.resourceType === 'db'
-  // ));
-  // console.log(resourceOptions);
-    // ({ data: resourceOptions['ACLB'] } = useLoadBalancerResourcesQuery(
-    //   props.resourceType === 'ACLB'
-    // ))
 
   React.useEffect(() => {
     formik.setFieldValue(
@@ -81,11 +66,6 @@ export const CloudViewMultiResourceSelect = (
         return option.id === value.id;
       }}
       onChange={(_: any, resources: any) => {
-        // console.log(resources.map((resource: any) => resource.id));
-        // formik.setFieldValue(
-        //   `${props.name}`,
-        //   resources.map((resource: any) => resource)
-        // );
         setResource(resources);
       }}
       autoHighlight
